+++ Gemfile
@@ -6,6 +6,7 @@ gem 'data_mapper'
 gem 'dm-postgres-adapter'
 gem 'bcrypt'
 gem 'sinatra-flash'
+gem 'sinatra-partial'
 gem 'pry-byebug'
 
 group :test do
