+++ app/views/_footer.erb
@@ -1,0 +1,3 @@
+<footer class="footer">
+  <p>&copy; 2016, Arunas Skirius</p>
+</footer>
